SRE-7111 - Add a second handler to handle SNS events
The S3Event is wrapped with an SNS event wrapper object when it is
delivered to the lambda to pushes events to elastic search. The entire
configuration and the code is almost the same as the handler for
S3Event, except that the ElastisLambdaSNSForwarder unpacks the received
SNSEvent to extract the S3Event.